♻️ Move some type schemas out into schemas/ to prevent circular imports
